add most optimal AC version of problem 6

+++ First Cpp contest/problem_6.cpp
@@ -6,10 +6,11 @@ int main(){
     while(t--){
         int N;
         cin >> N;
-        vector<int> prev_v = {1};
+        vector<long long> prev_v = {1};
 
         for(int i=0; i < N; i++){
-            vector<int> current_v = {1};
+            vector<long long> current_v = {1};
+
             for(int k = 1; k < i; k++){
                 current_v.push_back(prev_v[k] + prev_v[k-1]);
             }
@@ -17,11 +18,10 @@ int main(){
             if(i > 0){
                 current_v.push_back(1);
             }
-            for (int x : current_v){
+            for (long long x : current_v){
                 cout << x << " ";
             }
                 
-
             prev_v = current_v;
             cout << endl;
         }
